A first model of the map
Added a Location-class which can either take a string or longitud/latitud as input. Converts and stores as String either way. This could be an issue, maybe we want to display latitud/longitud sometimes? This however is more convenient for the URLBuilder.
The URLBuilder takes some input (eg where markers should be etc) and returns an url to a static map image. For more specific features, read the code (mostly the field declarations) or ask me (Florian) in person. Another way of doing this would be to type
https://developers.google.com/maps/documentation/staticmaps/
into your browser.
The MapModel class contains ALL data related to the map properties. The URLBuilder takes some of this as input and returns a link.

This is just the first iteration. More work has to be done.