Can calculate pixel-coordinates of the markers on a map
Location: added two properties, latitude and longitude. They are neccesary for the PixelFinder to work correctly. Each location now has a name, a latitude and a longitude.
Main: Just updated a simple test
PixelFinder: calculates the pixel coordinates of a list of locations, and the center of the image. This makes it possible to place interactive markers upon the map, since we then know exactly where to put them.