XMLParser extracts certain elements + fixed imports
Imports: Fixed some of the unused ones, since they gave me warnings.

XMLParser: Can now extract certain elements from an XML file (the attribute name has to be specified in a list). The class right now is full of uncommented code which has to be removed, it is far away from modular, and it gives output to a output stream instead of returning some kind of string or other resource. ie there is still a lot of work to do, though it should be possible to make a quick change if one want to implement it in the project.