A first working version of the mapPanel
Note: There is no such thing as structure, neither in the framwork or in the main-class. I just wanted to get a quick working version, that's all. It really is a lot of structuring/refactoring to do (split up into classes, remove unused code etc)

Capabilities: If queried in the proper manner (which is in an unlogical way right now, unless you don't call the main-method) it will print out the pixel coordinates of the locations on the map. Nice, huh? But still - the code is somewhat of a BIG mess.