Cleaning the code, fixed TODO's, added exception handling
LocationObject - added exceptionHandling for the case of a badly formatted string.
MapPanel - comments and TODO's cleanup, decommented away a tileserver which always gave an error before and changed the mouse-cursor type when on the map.
MapWrapper - added exception-handling for the unprobable case of 0 locations. Javadoc on most methods. Changed privacy from public to private on most methods.
Metadata - cleaned up some comments.