Implemented a working map-zoom
LocationMarker - implements PropertyChangeListener so it can listen to when the map zooms or updates it's position. Also added functions to handle these type of events (ie proper positioning after zoom or drag).
MapPanel - removed the Observer-pattern and let it fire a property change when it zooms or updates it's mapPosition. Note that the inner class MapState now is kind of unneccesary (doesn't work as a Observable anymore), so it should be more appropriate to put it's variables in the mapPanel and delete the class. However, I will look to do this soon.
MapWrapper - no longer implements Observer. Also stores the locationMarkers as a variable, so it does not have to reconstruct them every time the map gets dragged or zoomed.
Test - changed the size of the window
Images - added two possible marker-images.
.gitignore - added db.lck, log/log.ctrl, log/logmirror.ctrl and derby.log to this file. BUT the whole map still lies in the gitignore, so the other files won't be pushed neither way. We really need to talk it over which files/directorys should be pushed, and which not. (ie seg0?)