Starting to implement a markerController
The listener will be an inner class in the abstractJOverlayLabel (probably). That way it can access the pictures from the markers, if it wants to display a panel with info. The listener gets assigned in the controller, so when it gets clicked, it updates the view. Well, no code actually written yet, but I am on my way.