Initial geocoding - working system.out.println
Uses the nominatim geocoding service, which is limited to 1 search request per second. No big deal though.
Code can: create a legit URL-request
parse/extract the needed information from the XML-file (from the URL)
Print it out (currently with System.out.println)

Renamed package to geocoding, since that is what it does.

Todo: Integrate more smoothly, correct output source (not the system) and make a thread run it to limit search-requests to a maximum of 1 per second or so.