Cleanup in the mapmodel
Removed unused methods, referred with this to instance variables, made class use getter to access instance variables.
Also made the AbstractComponentModel more general, by renaming some of the functions and one field. Instead of images, it now holds locations (ok, this clearly doesn't sound so smart, but take a look at the code if that's the case or ask me in person, I'm just a little bit to lazy to describe it in words right now).